MQTT actions return the sequence ID.
The sequence ID is needed when waiting for the result of the command.
Now every action that uses one will return it.

Thanks to the above, the CLI can wait for the result of commands like
send-g-code.

The MQTT server only supports QoS 0 and the paho library blocks in
different places if I try to use 1 or 2 (specially 2), so we're back to
QoS 0 (fire-and-forget).

Signed-off-by: Gonzalo Paniagua Javier <gonzalo.mono@gmail.com>